Increment Script Run Count on Fragment Run Finished in Addition to Run Finished (#10012)
## Describe your changes

There is a bug in the forward message cache where the count of script
runs becomes out of sync between the frontend and the backend due to
differences in how fragment runs are counted. The result is that
sometimes messages that the backend believes are cached, are not
available to the frontend. There is an endpoint `_stcore/message` that
the frontend uses to request the missing message so when running
Streamlit in the typical way using the tornado server, the effect of the
bug is minimal, an increase in request to `_stcore/message` will be
observed, but the app will continue to function.

We are planning a larger assessment of the forward message cache with
the goal of simplification, but this change fixes the discrepancy in the
shorter term.

The tradeoff is that in apps using `st.fragment` cached messages may
expire more quickly. This can be controlled by the developer by raising
the `global.maxCachedMessageAge` value to compensate.

This change will affect the value of `script_run_count` which is stored
on `SessionInfo`. I have done a search of the codebase to confirm this
value is used only for managing the age of messages in the forward
message cache so we are not expecting unknown side effect to other
functionality.

This code is gated by a hidden config
(`global.includeFragmentRunsInForwardMessageCacheCount`) with a default
of False since we are not ready to release it broadly.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.