Add support for borders to audio & file input (#10223)
## Describe your changes

This PR adds a border to audio_input & file_uploader if
`widgetBorderColor` is set:


![image](https://github.com/user-attachments/assets/b6f51ec3-6895-4baf-9a6a-1b10519f194a)

`widgetBorderColor` is currently only used in SiS, but it's expected
that all input fields should have this border.

## Testing Plan

- The advanced theme used by SiS is currently only tested in a very
minimal way. We plan to add tests for advanced theming, but this will be
a dedicated follow up project that involves more work.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.