Force react-syntax-highlighter to cjs build (#10244)
## Describe your changes

There are issues with react-syntax-highlighter as an ESM module with
Vite. As a result, we explicitly choose the common js format in order to
fix the issue.

The steps here
https://github.com/react-syntax-highlighter/react-syntax-highlighter/issues/565
help illustrate the issue

## GitHub Issue Link (if applicable)
closes #10231

## Testing Plan

- Automated testing should still test everything works properly
- Manual testing to attempt to reproduce the issue.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.