Migrate TypeScript Configuration to its own package (#10188)
## Describe your changes

Refactor the TypeScript Configuration into a separate package and re-use
it in both app and lib.

While working on this, I helped resolved some type issues that soon
arose (Not sure what led to them all of a sudden failing, but they
seemed valid things to fix).

The build for lib was not completely building the library. I modified
the vite config to build the library properly with types and all. Things
seemed to look good.

Note: The fact that we use @streamlit/lib/src causes confusion between
the node module and the alias. Specifically eslint was failing with
regard to private variables. I removed it to make this PR easy. I modify
the name in #10213 and re-introduce the private variables. I can merge
it first, so it's all clean.

## Testing Plan

- All tests should pass.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.