add ability to reload page on version mismatch based on hardcoded client version (#10232)
## Describe your changes

- Changes the `hasStreamlitVersionChanged` logic to compare the version
between the `initializeMessage` and the variable
`PACKAGE_METADATA.version`. `PACKAGE_METADATA.version` is a global
variable that gets set by vite and is read from the `version` field in
streamlit/app's package.json.
- The previous logic relied on having sent a prior NewSession message so
that `sessionInfo.isSet` is true when the subsequent NewSession message
comes in. The new logic should suffice and remove the need to send
multiple NewSession messages to detect a version mismatch. The frontend
knows what version it is based on what's in package.json, which gets
revved for every streamlit release.
- SiS currently sends an initialization sequence of NewSession, Alert
FwdMessage pairs. We are removing this initialization sequence so that
we can show the built-in streamlit skeleton during loading. SiS will
render the loading sequence outside of streamlit as a page footer.
Therefore, the current logic that assumes the host will send multiple
NewSession messages will no longer be correct.
- This new behavior is only enabled if
`window.__streamlit?.ENABLE_RELOAD_BASED_ON_HARDCODED_STREAMLIT_VERSION`
is set to `true`, allowing it to be enabled behind a feature flag.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python) ✅ 
- E2E Tests
- Any manual testing needed? ✅

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.