Update file uploader to uses new `accept` value (#10288)
## Describe your changes

The `accept` props changed to accepting object of mimetypes as keys in
the new `react-dropzone` dependency.
This would break existing widgets that uses the `type` param in
`st.file_uploader` to define accepted extensions.

Here we added a custom mimetype `application/streamlit` as a wild card
type to still allow us to enforce the accepted extensions to satisfy the
version upgrade, with #10068 being the work in progress to support
official mimetypes.


## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python)
- E2E Tests
- Also manually tested with `st_file_uploader.py` on a couple file types
(txt, png, mov)

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.