Utilize new protobuf package in @streamlit/app and remove the export from @streamlit/lib (#10294)
## Describe your changes

In the effort to decouple a connection package from @streamlit/lib, we
are looking for easy ways to consolidate imports so that the dependency
tree is simpler (and lib is less the keeper of all things).

Now is the time to remove the protobuf export in lib (as that's now
handled in @streamlit/protobuf). This change removes that and updates
all of the references in @streamlit/app

## Testing Plan

- Tests should pass and builds should succeed

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.