Replace logging with loglevel (#10287)
## Describe your changes

In the effort to decouple a connection package from @streamlit/lib, we
are looking for easy ways to consolidate imports so that the dependency
tree is simpler (and lib is less the keeper of all things).

One of the biggest is the logging. It's used in many places, and it
simply is just passthrough to console. This PR replaces the logging with
the [loglevel](https://github.com/pimterry/loglevel) library. A very
light weight library with some logging features.

## Testing Plan

- tests should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.