Decouple SessionInfo from connection (#10297)
## Describe your changes
In the effort to decouple a connection package from @streamlit/lib, we
are looking for easy ways to consolidate imports so that the dependency
tree is simpler (and lib is less the keeper of all things).

Right now the connection relies on SessionInfo for one thing, getting
the last sessionId. This change converts it to a function that can be
provided easily.

Small, but I also refactored out `ensureError`. It's one of those,
"Sure, it's convenient, but do we really need it?" This was designed
mostly to ensure typescript compatibility, so incorrect usage will be
flagged and the conditional is just as easy.

## Testing Plan

- Tests should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.