Remove BaseUriParts type in favor of standard Location (#10293)
## Describe your changes

In the effort to decouple a connection package from @streamlit/lib, we
are looking for easy ways to consolidate imports so that the dependency
tree is simpler (and lib is less the keeper of all things).

Here BaseUriParts is a subset of URL in properties. The key difference
is that the pathname by default has a `"/"` always in front. While it
can be helpful, I believe it's better to leverage the URL
type/convenience and just adjust the code for these differences.

## Testing Plan

- Tests should pass.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.