Remove PerformanceEvents (#10298)
## Describe your changes

PerformanceEvents is great, but we have since migrated to Browser
features. This change was mainly used to test `fastReruns`, which is no
longer necessary (On top of that, it seems like we should have a package
for the logging, some third-party library, or a separate small package
for this use case.

## Testing Plan

- Tests should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.