Remove React from connection logic (#10301)
## Describe your changes

In the effort to decouple a connection package from @streamlit/lib, we
are looking for easy ways to consolidate imports so that the dependency
tree is simpler.

There's no real reason to have React be included in such a package. The
right way to handle this is likely to use error codes, but a simple
setup for now is to migrate to Markdown.

## Testing Plan

- Verified React components for errors are the same.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.