Add `theme.linkColor` theming config option (#10271)
## Describe your changes

Adds the `theme.linkColor` config option that allows to configure the
color of links in Streamlit.

This PR also adds a secondary color, which is only used internally for
now and not exposed to the user.

## GitHub Issue Link (if applicable)

- Related to https://github.com/streamlit/streamlit/issues/8844

## Testing Plan

- Adds unit & e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.