Remove `rich` from required dependencies (#10320)
## Describe your changes

We have a Python dependency in Streamlit - `rich` - that is almost
exclusively used in Community Cloud to get nicer formatting for
exceptions in the log viewer. However, it is currently installed for all
Streamlit installs, which is suboptimal. The dependency is already
implemented in Streamlit in a way that it is optional (only used if
installed). We have recently changed that `rich` gets auto-installed on
Community Cloud, therefore, we can remove it from Streamlit's required
dependencies.

In a follow-up after the release, we also plan to deprecate and
eventually remove the `logger.enableRich` config option and always use
`rich` if installed.

## Testing Plan

- No logical changes. Rich is already an optional dependency in the code
logic.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.