Deprecate widget and skeleton background color options (#10332)
## Describe your changes

It deprecates the `widgetBackgroundColor` and `skeletonBackgroundColor`
(internal) theme options. I have verified that SiS does not use these
options or uses it with the same values that are also provided for
alternative options widgetBackgroundColor -> secondaryBackgroundColor.

## Testing Plan

- Only removes logic. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.