Refactor & simplify sidenav theming (#10328)
## Describe your changes

Refactors and simplifies the colors used in sidebar nav to only use
colors adapted from the text color to work correctly across the theme.
This changes causes very minor differences in look which are not even
detected by our snapshot matching.

## Testing Plan

- No logical and only very minor visual changes. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.