Fix issue with active pills & segmented control label font-size (#10349)
## Describe your changes

Configure usage of the smaller font size directly with Markdown instead
of applying CSS from outside, since in some browsers it might lead to
weird resizing:


https://github.com/user-attachments/assets/da0b48c8-b485-40b0-b12e-308bef6a48f6

## Testing Plan

- I wasn't able to reproduce this in most of the common browsers, only
saw this happening in the playwright browser which seems to have a
slightly different priority of how styles are applied.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.