Allow configuring the row height for `st.dataframe` and `st.data_editor` (#9549)
## Describe your changes

Adds a `row_height` parameter to `st.dataframe` and `st.data_editor`
that allows the configuration of the row height. Also, potentially
auto-activated word wrapping in text columns as shown here:

- Prototype: https://configurable-row-height.streamlit.app


![image](https://github.com/user-attachments/assets/fc176769-97b1-47d1-bc33-e842c81e3515)

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/7266
- Closes https://github.com/streamlit/streamlit/issues/8632
- Closes https://github.com/streamlit/streamlit/issues/5386

## Testing Plan

- Added e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.