Increase heap size for lib build (#10418)
## Describe your changes

Our tests intermittently fail due to heap space on the lib package.
Let's increase the size to make sure it doesn't happen.

NOTE: I think this will be less necessary as we continue to break
packages out thoughtfully. The build steps will be simpler and the
imports will be less unwieldy.

## Testing Plan

- Tests should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.