[Fix]: `st.altair_chart` selection interval showing `true` tooltip (#10456)
When adding a selection interval to an altair plot, the selection rectangle has a tooltip "true". There should not be a tooltip for the selection mark - it should be the same as vega behavior.

As noted in the vega docs [here](https://vega.github.io/vega-lite/docs/mark.html#mark-def) for `tooltip` property, setting tooltip to `true` or `{"content": "encoding"}` should exhibit the same behavior.

GitHub Issue Link - Closes #10448